OCR/requirements.txt
Cython==0.29.28
install==1.3.5
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.0
pyparsing==3.0.8
pytesseract==0.3.9
tesseract==0.1.3